+++ resources/recipes/configure.rb
@@ -6,7 +6,11 @@
 # Services configuration
 
 # manager services
-manager_services = manager_services()
+managers = node.run_state['managers']
+manager_services = node.run_state['manager_services']
+node.default['redborder']['manager']['services']['current'] = node.run_state['manager_services']
+virtual_ips = node.run_state['virtual_ips']
+virtual_ips_per_ip = node.run_state['virtual_ips_per_ip']
 
 rb_common_config 'Configure common' do
   action :configure
@@ -44,6 +48,31 @@ chef_server_config 'Configure chef services' do
     action [:add, :register]
   else
     action [:remove, :deregister]
+  end
+end
+
+begin
+  vrrp_secrets = data_bag_item('passwords', 'vrrp')
+rescue
+  vrrp_secrets = {}
+end
+
+keepalived_config 'Configure keepalived' do
+  vrrp_secrets vrrp_secrets
+  virtual_ips virtual_ips
+  virtual_ips_per_ip virtual_ips_per_ip
+  managers managers
+  balanced_services node['redborder']['manager']['balanced']
+  has_any_virtual_ip node.run_state['has_any_virtual_ip']
+  manager_services manager_services
+  ipmgt node['ipaddress']
+  iface_management node['redborder']['management_interface']
+  ipaddress_sync node['ipaddress_sync']
+  managers_per_service node['redborder']['managers_per_services']
+  if manager_services['keepalived']
+    action :add
+  else
+    action :remove
   end
 end
 
@@ -522,7 +551,11 @@ minio_config 'Configure S3 (minio)' do
   ipaddress node['ipaddress_sync']
   access_key_id s3_secrets['s3_access_key_id']
   secret_key_id s3_secrets['s3_secret_key_id']
-  action((manager_services['s3'] && (external_services['s3'] == 'onpremise')) ? [:add, :register] : [:remove, :deregister])
+  if manager_services['s3'] && (external_services['s3'] == 'onpremise')
+    action [:add, :register]
+  else
+    action [:remove, :deregister]
+  end
 end
 
 # First configure the cert for the service before configuring nginx
@@ -537,7 +570,9 @@ end
 # Configure Nginx s3 onpremise nodes for now..
 minio_config 'Configure Nginx S3 (minio)' do
   s3_hosts node['redborder']['s3']['s3_hosts']
-  action((manager_services['s3'] && (external_services['s3'] == 'onpremise')) ? [:add_s3_conf_nginx] : :nothing)
+  if manager_services['s3'] && (external_services['s3'] == 'onpremise')
+    action [:add_s3_conf_nginx]
+  end
 end
 
 begin